Move Window to pygame.window.Window namespace
