+++ buildconfig/Setup.Android.SDL2.in
@@ -61,5 +61,5 @@ pixelarray src_c/pixelarray.c $(SDL) $(DEBUG)
 math src_c/math.c $(SDL) $(DEBUG)
 pixelcopy src_c/pixelcopy.c $(SDL) $(DEBUG)
 newbuffer src_c/newbuffer.c $(SDL) $(DEBUG)
-_window src_c/window.c $(SDL) $(DEBUG)
+window src_c/window.c $(SDL) $(DEBUG)
 geometry src_c/geometry.c $(SDL) $(DEBUG)
