+++ buildconfig/Setup.SDL2.in
@@ -73,4 +73,4 @@ pixelcopy src_c/pixelcopy.c $(SDL) $(DEBUG)
 newbuffer src_c/newbuffer.c $(SDL) $(DEBUG)
 system src_c/system.c $(SDL) $(DEBUG)
 geometry src_c/geometry.c $(SDL) $(DEBUG)
-_window src_c/window.c $(SDL) $(DEBUG)
+window src_c/window.c $(SDL) $(DEBUG)
