+++ buildconfig/stubs/pygame/window.pyi
