+++ src_c/static.c
@@ -176,7 +176,7 @@ PyMODINIT_FUNC
 PyInit_pixelarray(void);
 
 PyMODINIT_FUNC
-PyInit__window(void);
+PyInit_window(void);
 
 // pygame_static module
 
@@ -306,7 +306,7 @@ PyInit_pygame_static()
     load_submodule("pygame", PyInit_pg_mixer(), "mixer");
     load_submodule("pygame.mixer", PyInit_mixer_music(), "music");
 
-    load_submodule("pygame", PyInit__window(), "_window");
+    load_submodule("pygame", PyInit_window(), "window");
 
     load_submodule("pygame", PyInit_pixelarray(), "pixelarray");
